+++ hosts/delta/home.nix
@@ -1,1 +1,5 @@
-{ ... }: { }
+{ pkgs, ... }: {
+  home.packages = builtins.attrValues {
+    inherit (pkgs) discord;
+  };
+}