+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Added `discord` on `delta`
 - Added `aria2`
 - Disabled `VSCode` automatically checking for updates as the version is managed by `nixpkgs`
 - Updated `ranger` patch to work when `TERM_PROGRAM` (from `xterm.js`) is not set